requirements.txt
requests
bs4
lxml
python-dateutil
